+++ test/sample-sync-doc-definitions-spec.js
@@ -459,16 +459,33 @@ describe('The sample-sync-doc-definitions sync function', function() {
     var expectedDocType = 'notification';
     var expectedBasePrivilege = 'NOTIFICATIONS';
 
+    function getExpectedAccessAssignments(doc, docId) {
+      return [
+        {
+          expectedChannels: [ docId + '-VIEW' ],
+          expectedUsers: doc ? doc.users : null,
+          expectedRoles: doc ? doc.groups : null
+        }
+      ];
+    }
+
     function verifyNotificationCreated(businessId, doc) {
-      testHelper.verifyDocumentCreated(doc, serviceChannel);
+      testHelper.verifyDocumentCreated(doc, serviceChannel, getExpectedAccessAssignments(doc, doc._id));
     }
 
     function verifyNotificationReplaced(businessId, doc, oldDoc) {
-      testHelper.verifyDocumentReplaced(doc, oldDoc, [ serviceChannel, businessId + '-CHANGE_' + expectedBasePrivilege ]);
+      testHelper.verifyDocumentReplaced(
+        doc,
+        oldDoc,
+        [ serviceChannel, businessId + '-CHANGE_' + expectedBasePrivilege ],
+        getExpectedAccessAssignments(doc, doc._id));
     }
 
     function verifyNotificationDeleted(businessId, oldDoc) {
-      testHelper.verifyDocumentDeleted(oldDoc, [ serviceChannel, businessId + '-REMOVE_' + expectedBasePrivilege ]);
+      testHelper.verifyDocumentDeleted(
+        oldDoc,
+        [ serviceChannel, businessId + '-REMOVE_' + expectedBasePrivilege ],
+        getExpectedAccessAssignments({ }, oldDoc._id));
     }
 
     function verifyNotificationNotCreated(businessId, doc, expectedErrorMessages) {
@@ -492,7 +509,8 @@ describe('The sample-sync-doc-definitions sync function', function() {
         subject: 'pay up!',
         message: 'you best pay up now, or else...',
         createdAt: '2016-02-29T17:13:43.666Z',
-        actions: [ { url: 'http://foobar.baz', label: 'pay up here'} ]
+        actions: [ { url: 'http://foobar.baz', label: 'pay up here'} ],
+        users: [ 'foobar', 'baz' ]
       };
 
       verifyNotificationCreated(63, doc);
@@ -535,7 +553,8 @@ describe('The sample-sync-doc-definitions sync function', function() {
         message: 'last warning!',
         createdAt: '2016-02-29T17:13:43.666Z',
         firstReadAt: '2016-07-14T21:21:21.212-08:00',
-        actions: [ { url: 'http://foobar.baz/lastwarning', label: 'pay up here'} ]
+        actions: [ { url: 'http://foobar.baz/lastwarning', label: 'pay up here'} ],
+        groups: [ 'my-group1', 'my-group2' ]
       };
       var oldDoc = {
         _id: 'biz.7.notification.3',
@@ -544,7 +563,8 @@ describe('The sample-sync-doc-definitions sync function', function() {
         subject: 'a different subject',
         message: 'last warning!',
         createdAt: '2016-02-29T17:13:43.666Z',
-        actions: [ { 'url': 'http://foobar.baz/lastwarning', 'label': 'pay up here'} ]
+        actions: [ { 'url': 'http://foobar.baz/lastwarning', 'label': 'pay up here'} ],
+        groups: [ 'my-group1', 'my-group2' ]
       };
 
       verifyNotificationReplaced(7, doc, oldDoc);